Refatora pastas e apaga components desnecessários
