+++ src/app/auth/cadastro/cadastro.component.html
@@ -48,5 +48,5 @@
         </form>
         <p class="textoLogin">Já tem login? Clique <a class="link" href="/login">aqui.</a></p>
     </div>
-    <Rodape></Rodape>
+    <footer></footer>
 </div>
