--- src/app/home/conteudo/conteudo.component.css
@@ -1,189 +1,0 @@
-#logo {
-    color: #fff;
-}
-
-#logo>a:visited,
-a:hover,
-a:active,
-a:link {
-    text-decoration: none;
-    color: white;
-}
-.is-invalid{
-    color: red;
-}
-
-.sidenav-container {
-    height: 100%;
-}
-
-.sidenav {
-    width: 200px;
-    background-color: #456c45;
-}
-
-.sidenav .mat-toolbar {
-    background: #5b8f6f;
-}
-
-.toolbar-home {
-    top: 0;
-    position: sticky;
-    display: flex;
-    justify-content: space-between;
-    z-index: 1;
-    padding: 0 8vw;
-    background-color: #406040;
-}
-
-.logo {
-    font-family: Montserrat;
-    font-weight: 800;
-    font-size: 2em;
-    text-decoration: none;
-    color: white;
-    background-image: url(../../../assets/images/logo-white.png);
-}
-
-.logo-img {
-    width: 180px;
-    padding-bottom: 4px;
-}
-
-.btn-login {
-    font-family: "Barlow", sans-serif;
-    background: transparent;
-    color: #fff;
-    border: 2px solid white;
-    padding: 6px 16px;
-    border-radius: 4px;
-    font-weight: 500;
-    font-size: 1.3rem;
-    cursor: pointer;
-}
-
-.btn-wrapper {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    padding: 0 20px;
-    margin: 32px 0;
-}
-
-.btn-home {
-    font-family: "Barlow", sans-serif;
-    background: #406040;
-    color: #fff;
-    width: 320px;
-    height: 50px;
-    margin: 10px 24px;
-    text-decoration: none;
-    border: none;
-    padding: 0 20px;
-    border-radius: 4px;
-    transition: 100ms;
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    position: relative;
-    overflow: hidden;
-    cursor: pointer;
-    font-weight: 500;
-    font-size: 1.3rem;
-    line-height: 1.875rem;
-}
-
-.item1 {
-    top: 0;
-    display: flex;
-    margin: 0 auto;
-    flex-direction: column;
-    justify-content: center;
-    flex-wrap: wrap;
-    text-align: center;
-    width: 100%;
-    background-image: url(src/assets/images/fundo-home-frutas.jpeg);
-}
-
-.title-item1 {
-    color: #406040;
-    font-size: 1.8em;
-    margin: auto;
-    padding-right: 15px;
-    padding-left: 15px;
-}
-
-.text-item1 {
-    color: #406040;
-    padding: 5px 15px;
-    font-size: 1.4em;
-}
-
-.title-item {
-    font-size: 1.8em;
-    margin: auto;
-    text-align: center;
-    padding-right: 15px;
-    padding-left: 15px;
-    margin-bottom: 64px;
-}
-
-.text-item {
-    padding: 5px 15px;
-    font-size: 1.4em;
-    text-align: center;
-}
-
-.icon-item {
-    font-size: 4em;
-}
-
-.item2 {
-    flex-wrap: wrap;
-    text-align: center;
-}
-
-.item3 {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    flex-wrap: wrap;
-    text-align: center;
-}
-
-.item3-section {
-    margin: 3%;
-    width: 300px;
-}
-
-footer {
-    text-align: center;
-}
-
-.form {
-    display: flex;
-    flex-direction: column;
-    margin-top: 5%;
-    margin-left: 5%;
-    margin-right: 5%;
-}
-
-.button-form {
-    display: flex;
-    color: white;
-    background-color: #456c45;
-    margin: auto;
-}
-
-@media only screen and (max-width: 768px) {
-    .btn-wrapper {
-        padding: 0;
-        margin: 32px 0;
-    }
-    .btn-home {
-        width: 300px;
-    }
-    .btn-login {
-        display: none;
-    }
-}
