--- src/app/home/home.component.html
@@ -1,2 +1,0 @@
-<Conteudo></Conteudo>
-<Rodape></Rodape>
