--- src/app/home/home.component.ts
@@ -1,14 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'Home',
-  templateUrl: './home.component.html',
-  styleUrls: ['./home.component.css']
-})
-export class HomeComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() { }
-
-}
