--- src/app/layout/menu-nav/menu-nav.component.html
@@ -1,26 +1,0 @@
-<mat-sidenav-container class="sidenav-container">
-    <mat-sidenav #drawer class="sidenav" fixedInViewport [attr.role]="(isHandset$ | async) ? 'dialog' : 'navigation'" [mode]="(isHandset$ | async) ? 'over' : 'over'">
-        <mat-toolbar class="menu">Menu</mat-toolbar>
-        <mat-nav-list>
-            <a mat-list-item href="#item1" class="menu-item" (click)="drawer.toggle()">Acabe com a fome</a>
-            <a mat-list-item href="#item2" class="menu-item" (click)="drawer.toggle()">O que fazemos</a>
-            <a mat-list-item href="#item3" class="menu-item" (click)="drawer.toggle()">Como funciona?</a>
-            <a mat-list-item href="#item4" class="menu-item" (click)="drawer.toggle()">Sobre</a>
-            <a mat-list-item href="#item5" class="menu-item" (click)="drawer.toggle()">Contato</a>
-        </mat-nav-list>
-    </mat-sidenav>
-    <mat-sidenav-content>
-        <mat-toolbar color="primary">
-          <button type="button" aria-label="Toggle sidenav" mat-icon-button (click)="drawer.toggle()">
-              <mat-icon aria-label="Side nav toggle icon">menu</mat-icon>
-          </button>
-          <a href="/" class="logo">FoodCare</a>
-          <button mat-icon-button>
-              <a href="/login" class="button-login">
-                  <i class="material-icons">person_outline</i>
-              </a>
-          </button>
-        </mat-toolbar>
-        <ng-content></ng-content>
-    </mat-sidenav-content>
-</mat-sidenav-container>
