--- src/app/layout/menu-nav/menu-nav.component.ts
@@ -1,23 +1,0 @@
-import { Component } from '@angular/core';
-import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
-import { Observable } from 'rxjs';
-import { map, shareReplay } from 'rxjs/operators';
-
-@Component({
-  selector: 'MenuNav',
-  templateUrl: './menu-nav.component.html',
-  styleUrls: ['./menu-nav.component.css']
-})
-export class MenuNavComponent {
-
-  constructor(
-    private breakpointObserver: BreakpointObserver,
-  ) { }
-
-  isHandset$: Observable<boolean> = this.breakpointObserver.observe(Breakpoints.Handset)
-    .pipe(
-      map(result => result.matches),
-      shareReplay()
-    );
-
-}
