--- src/app/layout/rodape/rodape.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'Rodape',
-  templateUrl: './rodape.component.html',
-  styleUrls: ['./rodape.component.css']
-})
-export class RodapeComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
